requirements.txt
Flask
numpy
pandas
faiss-cpu
scikit-learn
keras
tensorflow